web/requirements.txt
flask
gunicorn
pandas
numpy
tqdm
requests
websocket-client
colorlog
cryptography
fake-useragent
flask-login
python-dotenv
